refactor(cli): Move runner functions to library
