+++ src/lib.rs
@@ -1,4 +1,115 @@
+use lazy_static::lazy_static;
+use mlua::chunk;
+use mlua::prelude::*;
+use std::path::PathBuf;
 #[cfg(feature = "cli")]
 pub mod cli;
 
 pub type Result<T> = anyhow::Result<T>;
+
+// These env vars are generated by autoconf *at compile time* in order to emulate the system
+// default Lua interpreter. If they exist we use them, else keep it to ourselves.
+lazy_static! {
+    #[derive(Debug)]
+    pub static ref LUA_PATH: &'static str = option_env!["LUA_PATH"].unwrap_or_else(|| "./?.lua");
+    #[derive(Debug)]
+    pub static ref LUA_CPATH: &'static str = option_env!["LUA_CPATH"].unwrap_or_else(|| "");
+}
+
+pub fn run(
+    input: Option<PathBuf>,
+    backend: Option<String>,
+    class: Option<String>,
+    debug: Option<Vec<String>>,
+    evaluate: Option<Vec<String>>,
+    evaluate_after: Option<Vec<String>>,
+    fontmanager: Option<String>,
+    makedeps: Option<PathBuf>,
+    output: Option<PathBuf>,
+    options: Option<Vec<String>>,
+    preamble: Option<Vec<PathBuf>>,
+    postamble: Option<Vec<PathBuf>>,
+    r#use: Option<Vec<String>>,
+    quiet: bool,
+    traceback: bool,
+) -> crate::Result<()> {
+    let lua = unsafe { Lua::unsafe_new() };
+    let lua_path: LuaString = lua.create_string(&LUA_PATH.clone()).unwrap();
+    let lua_cpath: LuaString = lua.create_string(&LUA_CPATH.clone()).unwrap();
+    let sile: LuaTable = lua
+        .load(chunk! {
+            package.path = $lua_path
+            package.cpath = $lua_cpath
+            local executable = require("core.pathsetup")()
+            return require("core.sile")
+        })
+        .eval()?;
+    sile.set("traceback", traceback)?;
+    sile.set("quiet", quiet)?;
+    if let Some(flags) = debug {
+        let debug_flags: LuaTable = sile.get("debugFlags")?;
+        for flag in flags {
+            debug_flags.set(flag, true)?;
+        }
+    }
+    let full_version: String = sile.get("full_version")?;
+    let sile_input: LuaTable = sile.get("input")?;
+    if let Some(expressions) = evaluate {
+        sile_input.set("evaluates", expressions)?;
+    }
+    if let Some(expressions) = evaluate_after {
+        sile_input.set("evaluateAfters", expressions)?;
+    }
+    if let Some(backend) = backend {
+        sile.set("backend", backend)?;
+    }
+    if let Some(fontmanager) = fontmanager {
+        sile.set("fontmanager", fontmanager)?;
+    }
+    if let Some(class) = class {
+        sile_input.set("class", class)?;
+    }
+    if let Some(paths) = preamble {
+        sile_input.set("preambles", pbs_to_sts(paths))?;
+    }
+    if let Some(paths) = postamble {
+        sile_input.set("postamble", pbs_to_sts(paths))?;
+    }
+    if let Some(path) = makedeps {
+        sile_input.set("makedeps", pb_to_st(&path))?;
+    }
+    if let Some(options) = options {
+        sile_input.set("options", options)?;
+    }
+    if let Some(modules) = r#use {
+        sile_input.set("use", modules)?;
+    }
+    if !quiet {
+        eprintln!("{full_version} [Rust]");
+    }
+    let init: LuaFunction = sile.get("init")?;
+    init.call::<_, _>(())?;
+    if let Some(input) = input {
+        let input_filename: LuaString = lua.create_string(&pb_to_st(&input))?;
+        if let Some(path) = output {
+            sile.set("outputFilename", pb_to_st(&path))?;
+        }
+        sile_input.set("filename", input_filename)?;
+        let process_file: LuaFunction = sile.get("processFile")?;
+        process_file.call::<LuaString, ()>(sile_input.get("filename")?)?;
+        let finish: LuaFunction = sile.get("finish")?;
+        finish.call::<_, _>(())?;
+    } else {
+        let repl: LuaTable = sile.get("repl")?;
+        repl.call_method::<_, _, _>("enter", ())?;
+    }
+    Ok(())
+}
+
+fn pb_to_st(path: &PathBuf) -> String {
+    path.clone().into_os_string().into_string().unwrap()
+}
+
+fn pbs_to_sts(paths: Vec<PathBuf>) -> Vec<String> {
+    paths.iter().map(pb_to_st).collect()
+}
